角色id改为int

+++ fhs_module/fhs_basics/fhs_basics_service/src/main/java/com/fhs/basics/service/impl/UcenterMsRoleServiceImpl.java
@@ -137,7 +137,7 @@ public class UcenterMsRoleServiceImpl extends BaseServiceImpl<UcenterMsRoleVO, U
     @Override
     public int deleteById(Serializable primaryValue) {
         UcenterMsRoleVO role = new UcenterMsRoleVO();
-        role.setRoleId(ConverterUtils.toInt(primaryValue));
+        role.setRoleId(ConverterUtils.toLong(primaryValue));
         // 删除按钮信息
         boolean count = deleteButtons(role);
         if (count) {
